+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Upgrade minimum VS Code version to 1.99.
 - `@macro` supports multiple comma-separated options.
 - Add new `retained` macro option.
+- Update `macros` agent prompt to improve code generation quality.
 - Fix: TreeView example does not load in case-sensitive file-systems.
 
 ## 0.3.3